pyproject.toml
[project]
name = "demos"
version = "0.1.0"
description = ""
authors = [
    {name = "GeorgeHit",email = "noricerice@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "langchain>=0.3.27,<0.4.0",
    "langchain-core>=0.3.0,<0.4.0",
    "langchain-community>=0.3.0,<0.4.0",
    "aiofiles>=23.0.0",
    "aiohttp>=3.8.0",
    "dashscope (>=1.24.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
